requirements.txt
flask==3.1.0
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
Pillow==11.1.0
numpy==2.1.1
pandas==2.2.3
matplotlib==3.10.1
pyyaml==6.0.2
requests==2.32.3